requirements.txt
customtkinter
google-generativeai
Pillow
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
